refactor: improve encoder precision and enhance ListOverlay state management
- Expand encoder discrete steps from uint8_t to uint16_t for greater precision
- Add LVGL symbols font (Montserrat 12) to font registry
- Replace CHECKED state with FOCUSED state in ListOverlay for better semantics
- Add comprehensive text styling for all UI states (default, focused, pressed, disabled)
- Implement recursive state application for consistent child object styling

+++ src/adapter/input/encoder/EncoderController.cpp
@@ -23,7 +23,7 @@ void EncoderController::resetEncoderPosition(EncoderID encoderId, float normaliz
     }
 }
 
-void EncoderController::setDiscreteSteps(EncoderID encoderId, uint8_t steps) {
+void EncoderController::setDiscreteSteps(EncoderID encoderId, uint16_t steps) {
     Encoder* encoder = getEncoder(encoderId);
     if (encoder) {
         encoder->setDiscreteSteps(steps);
